gen-manifests: support globs in cl args
Add support for globs in command line arguments for distros, arches, and
images (image types).
Globs are matched against the valid list of values.  Any value from the
command line that doesn't match a valid value is printed in a warning.

This is useful as a shortcut for a group of values that share a common
trait, for example '-images "*iot*"' to generate all iot manifests, or
for all versions of a distro family, for example '-distros "fedora*"'.
It's especially useful for test scripts that might need to generate
manifests for all version of a distro but might not be updated when a
new one is added.