build: read and build a single config
Builds a single image with a required configuration like the ones in
./test/configs.
The artifact is saved at the target location in a subdirectory
constructed from the distribution name, the architecture, the image
type, and the name in the configuration.  The manifest is also stored
in the same directory as manifest.json.