test/generate-build-configs: check for previous successful builds
The script generates all manifests and compares the content hashes for
each manifest against the S3 bucket where successful builds are
uploaded.  If a manifest's checksum is found in the bucket, it will not
configure a build pipeline for it.

The script doesn't rely on the list-images command and instead relies on
gen-manifests to create all supported configurations.
The test requires that build configs in ./test/configs have filenames
that match the "name" field so it verifies this at the start.
This isn't a requirement for the configs in general, so in other places,
configs can have any name/filename, but for this test it's necessary.