cmd/build: resolve ostree commits
The original manifest generation for the build command followed the same
logic as the gen-manfiests command which creates a fake commit ID for
sources.  In the manifests this is needed to be able to create test
manifests without having an ostree remote to pull from.  Here we need to
resolve commit IDs properly because we're going to actually build the
manifest.