test: invert config map
Invert the config map to be keyed by paths to config files.  Each config
file is mapped to a list of image types, distros, and arches.

Empty lists mean it applies to everything (since it makes little sense
to define a config and not map it to anything).  Globs are also
supported.

The all-customizations config now applies to qcow2 again but to all
distros except RHEL 7.

Updated the description in the README.