dnf-json: weak-deps selection to go
`dnf-json` had its own logic for determining which transactions to solve
with weak dependencies enabled (only the first one). Move that logic
into the Go code to make it easier to change which package specs get
solved *with* weak deps (or without).

A `PackageSet` now contains an `InstallWeakDeps` property that is
propagated all the way down when depsolving.