test/configs: test oscap customizations
Since, OpenSCAP config is not available for some of the minor versions
of RHEL8 and RHEL9, create a separate config for all customizations
including OpenSCAP. This is to ensure that the OpenSCAP configurations
don't break other customizations.

PCI-DSS was the chosen profile for the OpenSCAP customizations since
it is a profile that is common between RHEL8, RHEL9, Centos & Fedora.