dnf: mark packages
Package marking is enabled per image type and propagated down to the OS
pipeline. The package marking information is output by `dnf-json` based
on what `dnf` gave as reasons during the depsolve transactions. This is
stored on the `rpmmd.PackageSpec`.

Note that we currently output 'user' for any 'group'-installed package.

There's also a bit of complication here; as `dnf-json` does chain
depsolving it marks all packages from the previous transaction for
`install`. This leads to all packages from the previous transaction
being marked as `user`-installed. To circumvent this we keep track of
the reasons a package was selected and the oldest one wins.

Another complication is that a package could be initially resolved as a
dependency but be user-requested in a follow up transaction. Some logic
is involved to allow upgrades from dependency and weak-dependency to
user. We also blanket-allow upgrading weak-dependency to dependency.

Package marking is enabled for all Fedora image types.