test: dynamic pipeline generator for ostree builds
New dynamic pipeline generator similar to test/generate-build-config but
for image types that require pulling an ostree commit.
The script filters out configs that do not specify a dependency.

The script first collects all dependency build configurations (for
example edge-container, empty.json) and generates their manifests to
calculate the manifest ID.  The manifest ID is used to find the
corresponding container in the gitlab registry, which should have
already been built.  For each config that specifies a dependency
(for example test/configs/edge-ostree-pull-empty.json) the container for
the dependency is pulled from the gitlab registry and launched in order
to generate manifests with ostree commit resolution enabled.

Each config with a dependency is rewritten with the URL for the
corresponding container.  Once the manifest generation is done, the
builds are filtered in the same was as they are in
test/generate-build-config: for any manifest ID not found in the build
cache, a job is added to the pipeline to build it.

The script does not yet write the gitlab pipeline file for the builds.

Note that this pipeline generator can only configure builds for the
architecture that it is running on since it needs to run containers for
a given architecture when generating manifests.