test/generate-build-config: last stage ID for manifest checksum
Instead of hashing the manifest contents, use the last stage ID computed
from osbuild --inspect.  This has the following advantages:
- It is not affected by formatting (indentation, newlines).
- It is not affected by content URLs when the content is the same.  For
  example, an ostree commit from a different URL will not affect the
  stage IDs if the content hash is the same.

This gets around an issue we were going to face when generating
manifests from local ostree containers when the IP address or port of
the local container changes.  It also lets us stop worrying about the
exact format of the manifest file when comparing.

For now, the build config generator checks for both filenames
(hash-based and stage-id-based) in the build cache from s3.  This is
temporary until all or most files in the build cache are updated to the
new filenames.  The script renames and syncs back with the new filename
any file it finds that matches the old way of computing the manifest
checksum.