test/configs: new configs for pulling ostree commits
A new "depends" key in the configs indicates that a build using it will
need to pull an ostree commit from a specific image type built with
another config.

Two new configs are added to be used with ostree-based images that pull
an ostree commit and deploy it as an image or embed it in an installer.

These configs include an example URL so they can also be used when
commit resolution is disabled (gen-manifests -commits=false).