resolver: Always return results sorted by Digest
The output of the resolver can be inconsistent. This causes problems
with the diff-manifest CI script because consecutive runs on the same
code may have different outputs, causing a false manifest diff error
from Schutzbot.

The tests was already sorting by Digest, so move that up into Finished()
and remove the sort from the test. Now the results will always be in a
consistent order.