test/generate-ostree-build-config: write the build configs
Write the build configurations for the dynamic pipeline.

The $PARENT_PIPELINE_ID will need to be set in the main .gitlab-ci.yml
so that the child pipeline can access artifacts from the parent
pipeline.

The child job configs aren't complete yet.  It still needs the command
to start the container for the build job and the config needs to be
written as an artifact for the child job to read.