test: install dependencies outside build scripts
Install dependencies in the test job definition and not the build script
itself.  This makes the build script distro agnostic, in case we switch
certain runners to run in the shell runner for simplicity.

The build configs for the first child pipeline were also changed so that
the script is an array of commands instead of a multiline script.