github: trigger gitlab CI independently
Trigger gitlab CI as a separate workflow and don't depend on the other
tests.  Moved the gitlab-ci-helper to the separate workflow file so it
just runs immediately.  This also simplifies the artifact sharing.

The action metadata parsing now reads from `pull_request` instead of
`workflow_run` since the action is triggered on PRs.

Note also that the action is not run on main, only on pull requests and
the merge queue.  Running on main isn't needed since the merge queue
head should be equivalent.