github: move gitlab helper to separate workflow
Move the github helper to a separate workflow on its own and make the
trigger-gitlab workflow depend on it only.  This makes gitlab CI start
without waiting for the rest of the tests.  The gitlab workflow needs to
be a workflow_run type job, otherwise it doesn't have access to secrets
(the gitlab ssh key).

The triggers are also changed so that the action is not run on main,
only on pull requests and the merge queue.  Running on main isn't needed
since the merge queue head should be equivalent.

The trigger-gitlab workflow is run on success of both Tests and GitLab
for now, because the GitLab workflow wont run until it's merged into
main.