distro/fedora: configurable ignition platform
The kernel command line argument ignition.platform.id was previously
always set to "metal".  Make it configurable on the raw image.  It's
still always set to "metal" when ignition is enabled for now.

The commit includes some refactoring of the ignition option handling.

Added a guard to panic if ignition is enabled but the platform isn't
set.