manifest: introduce FilePipeline interface
An interface for pipelines that produce a single file; typically an
image file, but can be any type.  This will make it convenient to have
the interface as a dependency to pipelines that require a single image
file as input (like xz and qcow2) without depending on a specific
implementation of a pipeline.