manifest: Manifest() and BuildPipeline() on common interface
Add getters for the manifest and build pipeline references to the common
Pipeline interface.  These are used in almost all pipeline constructors
and it will be useful to make them part of the interface so that we
don't always have to use concrete types to get to them.

Also rename GetManifest() and GetPlatform() functions on the Tree
interface to drop the Get (I think it's nicer).