manifest: simplify Pipeline constructors
Every pipeline that depends on another pipeline can read the manifest
(and sometimes extra options) off the dependent pipeline.  Simplify the
constructors for these cases.

This could also be done for the build pipeline as well, but technically
pipelines can use different build roots from their dependents and
although we never do this, it's best to reflect this in our design and
keep the support open for any future requirements.