gitlab: remove unused scripts and artifact storage
Remove all unused schutzbot scripts.

The way tests are run for this repo, the journal should have nothing
relevant or useful.
More generally, none of the artifacts that are pushed to s3, that we
carried over from osbuild-composer, are very relevant here.

If we introduce any artifacts that are sensitive and can't be stored in
the GitLab job artifacts, we can reintroduce the scripts.