ostree: skip resolving ref if it is a checksum
If a ref to be resolved is already a checksum, just copy it to the
checksum field and don't resolve.
This is only valid for parent refs.  Image refs that look like checksums
will be caught by the Validate() function.

Supporting checksums for parent refs means that users can specify a
parent ostree commit that is not the head of a branch.

Resolves COMPOSER-1976