fedora: partial cleanup of COI boot packages
Add the CoreOS Installer boot packages to the platform for the image
type, since for the simplified-installer the image being built is the
installer and the payload (ostree commit) is prebuilt.
I'm not entirely sure if this is how we should do this, so I wont make
the change in RHEL and CentOS and I added a note to come back to this.

The biosdevname is added conditionally when the option is set and is
required by the dracut stage.