osbuild: helper functions for curl and ostree sources
The inline and skopeo sources have constructors for the source itself
(New...Source()) and the items (New...Item()) as well as an AddItem()
method.  These are convenient and make source creation consistent and
easier to validate.  Add the same functions for curl and ostree sources.

For the curl source, the method is called AddPackage() because
technically the source supports downloading any type of file.