ostree_deployment: enable/disable services also using presets
Extend the `OSTreeDeployment` pipeline implementation to enable /
disable services also using systemd presets in case Ignition is being
used.

The pipeline already enables / disables services using the Systemd
stage, which calls `systemctl`, but on some systems (since Fedora 37),
Systemd will reset all enabled / disabled services based on system
presets on first boot. This is not a problem for our images in general,
since they contain fake machine-id, which makes systemd think that the
system is not booting for the first time. This is however not the case
when Ignition is being used, since it signals to systemd to treat the
first boot as an actual first boot.

Enabling systemd services using `systemctl` is still kept in the
pipeline, since the end result will be the same even if services are
enabled / disabled multiple times via different means.

Related to:
https://github.com/osbuild/osbuild-composer/pull/3634
https://fedoraproject.org/wiki/Changes/Preset_All_Systemd_Units_on_First_Boot

Signed-off-by: Tomáš Hozza <thozza@redhat.com>