minimal-raw/rhel: Fix the kernel command line
None of these options should be hard coded. The serial console
on a lot of devices is not ttyS0, (eg it's ttymxc0 on i.MX8) and
it breaks console output on connected displays. Some of these
options are x86 specific and we want to use the specific names
for the network interfaces so drop all those.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>