minimal-raw/rhel: Add WiFi support
Add WiFi support and initial-setup to create users. The image
should have basic WiFi support as a lot of the devices have
WiFi so being able to use it out of the box is important. The
majority of the WiFi firmware is currently in linux-firmware but
also add the most common Intel cards too.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>