minimal-raw/rhel: Add initial-setup
By default there's no user and the root account should be
locked so we need to provide a means for images that maybe
distributed to enable a user to create accounts to login
as we don't want to, by default, unlock the root account
and provide default credentials as that's a major security
issue. If a user doesn't add an account in a blueprint
they end up with an unusable image.

We add libxkbcommon as it's required by i-s but isn't
pulled in.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>