disk: partition: Allow specifying a starting offset
It's useful for a number of usecases to be able to
specify a starting offset for the first partition on
disk. This is needed for some arm images to allow
space for firmware, but it's also often needed for
virt images so VMs are optimally places on disk.

Fix an issue where the starting offset isn't set for
partition offsets.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>