minimal-raw/fedora: fixes for the image type
Add WiFi support and initial-setup to create users. The image
should have basic WiFi support as a lot of the devices have
WiFi so being able to use it out of the box is important. The
majority of the WiFi firmware is currently in linux-firmware but
also add the most common Intel cards too.

We need inital-setup to create users, set root password etc as
we don't want to be using default user/password as it's a major
security risk. We add libxkbcommon as i-s needs it.

We need to start default services to ensure the right things run
and we end up with things like firewalls and networks.

Finally we need to setup the firmware for the RPi.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>