distro: add a kickstart file to /root for minimal-raw images
Currently initial-setup requires a kickstart file in the /root directory
to work.  Adding a temporary file node to the image config for
minimal-raw images so that initial-setup runs and a user can be created
if necessary.  The minimal-raw image has no other mechanism for creating
a user if one wasn't specified at build time, so we need to be sure that
initial-setup runs.