fedora: fix default kernel command line
Fix the default Fedora kernel command line. It was advertised as the
default but the reality is the options were for cloud, so set a real
default as the option and update the cloud options to something
special for them. It's not truly correct as at least one of the
options is x86 specific but when booted on other architectures those
kernels will ignore it (and I honestly think it's ancient cargo
culted kernel command line options).

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>