minimal-raw/fedora: Add offset to partitions
A number of edge devices, in most cases pre-prod that may be on
the edge of proper certification, need to write the firmware to
the beginning of the storage, while this is far from ideal the
default 2048 blocks (1Mb) is generally not enough and it means
when the firmware is written out we will zap the EFI ESP
partition so let's start this with an offset of 8Mb for
these images to ensure we have enough space for those (hopefully
rare) use cases.

Also the Fedora image, due to support requirements around devices
like the Raspberry Pi 3 require the default partition tables
format to be msdos because the initial program loader in silicon
doesn't support loading the firmware off a GPT partition table.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>