test/generate-build-config: call the boot script
Add the call to the boot script after the call to build-image in the
generated pipeline.  The location of the image file is calculated from
the build configuration and the name of the last pipeline in the
manifest and the filename is taken to be the name of only file in the
directory.  If more than one file is found (or none), the function
raises with an error.