boot-aws: separate setup and teardown
setup: Upload the image and boot it.  Print the resource IDs that were
created during setup so they can be saved to a file.
teardown: Read the resource IDs from the file created by setup and clean
them up, just like before.

The teardown functionality is implied by the presence of the
'--tear-down' flag.