image: add InstallWeakDeps option to DiskImage
Support specifying the InstallWeakDeps option through the DiskImage
ImageKind, so that distros creating an image of that type can control
the weak dependency selection.

The option is a pointer to a boolean so that the nil value can fall back
to the default internal value defined by the pipeline (which can vary).