distro/rhel9: Make /boot 600 MiB big on RHEL 9.3+
Changes in xfsprogs in 9.3 caused XFS partitions have less available space
(due to more space reserved for the log section).

We discovered that this causes issues with kernel updates. When 3 kernels are
already installed and dnf wants to install a new one (and replace the oldest
one), /boot must contain 4 kernels for a brief period of time, which
doesn't fit on the small partition. This causes the transation test to fail,
and dnf refuses to perform such a transaction.

There's a workaround: Remove one of the kernels manually using `dnf remove`
before running an update.

This is however quite surprising and annoying, thus this commits bumps the
size of /boot to 600 MiB starting from 9.3 (and on CentOS Stream).