Fedora: add basic support for s390x and ppc64le
Add support for s390x and ppc64le architectures in the Fedora
distribution. For now, support building only a few images, which are
relatively easy to test manually to sanity check the support,
specifically the qcow2 and container images.

The base partition table for these architectures is taken from RHEL-9
distro, with the change of the default filesystem to be 'ext4' instead
of 'xfs', to make them consistent with partition tables for existing
Fedora architectures.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>