Platform/ppc64le: return boot packages only if intended for booting
Adjust the ppc64le platform to return boot and build packages related to
booting, only if the platform is configured to be intended for booting.
Otherwise, building e.g. a container image on ppc64le fails, due to the
image package set excluding some of the packages required by boot
packages.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>