test: print configurations built in CI run
In the check-build-coverage step at the end of the CI pipeline, also
collect all the build configurations that were built in the specific
pipeline (identified by the CI_COMMIT_SHA) in the build info and print
it in the log.

This makes it a tiny bit more convenient to check which images were
rebuilt and see if any images weren't rebuilt when they were meant to or
were rebuilt when they weren't expected to.