Add unit test for current MountpointPolicies
Add a simple unit test for the current definition of
`MountpointPolicies` to capture various scenarios of what is currently
allowed and what is not.

This will be helpful in the next steps when the policy gets relaxed and
reworked.

Delete distro unit-tests which were responsible for checking that some
specific mountpoints are not allowed by the mountpoint policy. The
newly added policy unit test tests the same thing, but on a single
place.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>