test: print configurations built in PR
The same way we print a list of image configurations built in a specific
CI run, let's also print a list of configurations built in the same PR
(identified by the CI_COMMIT_BRANCH).

Again, this makes it a bit more convenient to check which images were
rebuilt in a PR, but we should eventually use this information to inform
PR authors and reviewers of changes in image definitions.