gitlab: add an init job that initialises the cache
The new cache-init job pulls the build info from S3 and updates the
pipeline cache for all the next jobs in the CI pipeline.  The policy is
set to pull-push because caches are shared across pipelines and this can
speed up the cache initialisation.

The configure-generators cache policy is also set to 'pull-push' because
it will first update the local copy from S3 to get the information for
any new builds and then push the updated info to the gitlab CI cache for
the next pipeline.