osbuild: add ostree.deploy.container stage
Creates an org.osbuild.ostree.deploy.container stage.
Validates the options with a regular expression that matches the schema.
Takes a container as input.  Validates the length of the input
references to be exactly 1.

Also added tests for the validators for options and inputs.

See https://github.com/osbuild/osbuild/pull/1402