GenCopyFSTreeOptions: don't hardcode FS root mount name
The `GenCopyFSTreeOptions()` function made an assumption when generating
the org.osbuild.copy stage options, that the mount name for the
filesystem root is "root". While this worked so far, the method should
not rely on this behavior, since it would break if we change the
implementation of escaping mountpoint paths.

Modify the function to determine the mount name of the Entity based on
the mountpoint path while iterating over all partition table
entities. Specifically, use the mount name of the one with
mountpoint `/`.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>