osbuild/device: use systemd-escape like escaping for mountable paths
Previously, the function used to escape filesystem paths of mountable
entities, which are used as device names, could generate the same name
for different devices. Specifically, it returned `root` as the device
name for entity with `/` and `/root` mount point. This results in
invalid osbuild manifests being generated.

Rework the escaping function by implementing a similar escaping
algorithm as used by systemd-escape, which will ensure that we always
generate a unique device name for devices with different mountpoints,
while still generating the same device name for the same device (with
the same mountponit path).

Signed-off-by: Tomáš Hozza <thozza@redhat.com>