gen-manifest: add a shortcut option for providing single image config
The gen-manifest accepts an image config map via the `-config` option,
which is good for CI and testing purposes.

However, imagine a situation when:
- you are making changes to code in the repository which affects
  generated manifests,
- you want to use a custom image config to test these changes and
  generate manifests to build them using osbuild

In this case, one can filter out distro / arch / image type using the
`gen-manifest` CLI options, instead of creating the image config map.
Passing an ad-hoc image config to `gen-manifests` on the CLI to use for
all image types and filtering what gets generated using the CLI options
is IMO much nicer (or easier to use) for development, than the current
approach (which involves basically creating two JSON files and editing
one of them each time one wants to use a different image config).

Change the behavior of the original `-config` option to accept a
single image config which will be used for all image types / distros /
arches. Note that this will almost certainly fail for some images, but
the intention of this option is to be used for development and together
with filtering options such as `-distros`, `-arches` and `-image-types`.

The original option is now available under a new name `-config-map`,
which is better aligned with what it does.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>