manifest: only set one ostree commit in OSTreeDeployment
The pipeline either way only supported one ostree commit, enforced when
adding the resolved commits to the pipeline during serializeStart() and
by checking the array length during serialize().  To simplify the
requirement and avoid potential confusion when reading the struct
itself, replace the array on the pipeline struct definition with a
single commit spec (pointer) and keep the len == 1 check in the
serializeStart() function that, by its interface, must accept an array.

Also document the commit fields for OSTreeDeployment.  These are
private, so they wont show up in public API docs, but it's still useful
to have them documented in the code.