manifest: remove ignition options from OSTreeDeployment constructor
Only required properties should be specified through the constructor to
make sure they are not accidentally left empty or nil.  Optional
properties are added via public fields of the pipeline object.

Since the ignition platform is optional, change it to a public field.
Also, to simplify the interaction between the two ignition properties,
remove the boolean that signifies if ignition should be enabled and use
the IgnitionPlatform string instead.  Treat the empty string as the
disabled value.