distro/rhel9: add weak-dep dracut modules explicitly
For installers (both Anaconda and CoreOS) on all distros, we define a
couple of dracut modules that rely on packages that are pulled in as
weak dependencies of other packages in the installer package set.  Let's
define these explicitly instead.
For other packages that are required by other modules, let's still rely
on package "hard" dependencies that are already defined.

In the future, we should derive the required packages dynamically from
the dracut modules defined in the pipeline.