manifest: move coreos.no_persist_ip into image functions
Not all image types that use ignition want to define the kernel option.
Specifically, the (upcoming) CoreOS images wont want to enable this, so
let's move it into the image function and add it when the ignition
options are added.  It is conditioned on the distro version for now, so
we can't add it to the global image type kernel options.

This change has no effect on any of the current image types / manifests.