internal: add new interla/assertx/PanicsWithErrorRegexp
This adds a new internal package with `testify/assert` extensions.
It contains the new 'PanicsWithErrorRegexp()` helper that can
be removed once upstream https://github.com/stretchr/testify/issues/1304
is fixed (or we contributed this back).

It could not be added to `internal/test` as it would have created
cyclic dependencies when importing from `pkg/osbuild`.