pkg: test that "empty" RepoConfig marshal to "empty" json
This is a regression test for an issue found during the review
of PR#284 [0]. The issue was that
```
	ModuleHotfixes *bool    `json:"module_hotfixes"`
```
lacked the required `omitempty` which may cause issues
with dnfjson.

As a quick and easy way to prevent this in the future
this commit adds a small test that ensures that we notice
any missing `omitempty`.

Note that the `rpmmd` part is not strictly needed but it
feels more correct to do the check in both places that
use a repoConfig.

[0] https://github.com/osbuild/images/pull/284#discussion_r1414048979