distro/test_distro: adjust to the distrofactory approach
Adjust the test_distro implementation to provide the FactoryFunc in the
same way as the regular disto implementation. Consolidate the
implementation to ensure that the distro Name always contains a release
version. In addition, reduce the exported symbols to limit the unwanted
use of them in other packages.

Modify affected unit tests, which previously relied on some of the
exported symbols from the test_distro.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>