Make `reporegistry` public, so that it can be used by osbuild-composer
Some logic related to `reporegistry` will need to change due to
introduction of dot to separate major and minor release version in
distro names. Both, `images` and `osbuild-composer` have their own copy
of the code, which does not scale. In general, the `images` repo needs
to handle repositories to generate testing manifests, so it makes sense
to keep the authoritative copy of `reporegistry` in this repository and
then just use it in `osbuild-composer`.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>