Introduce `distroidparser` package for parsing distro ID strings
Factor out the distro-specific ID String parsers into a separate
functions in each distro implementation. This allows to introduce the
`distroidparser` to have a common and consistent way to parse distro
ID Strings, including the use of distro-specific parsers. This will be
crucial for backward compatibility of distro names without the dot
separator as used by APIs and distro configurations.

The parser needs to be a separate package, because adding it to the
`distro` package, where the `distro.ID` struct is defined, would result
in an import loop.

In addition adjust the logic of the RHEL parsing code a bit, to enable
returning more sensible errors from the parser function.

Cover the newly added code with unit tests.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>