pkg/distro: simplify remediation stage options
Generalise the creation of the remediation stage options and extract the
function to the `oscap` package. This function also slightly reduces the
tight coupling with the `tailoring` stage options. However, this stage
still depends on some of the config from the `tailoring` stage.

Jira: https://issues.redhat.com/browse/COMPOSER-1994