test: use osbuild commit for build test filtering
Record the osbuild commit ID from the Schutzfile in the build info when
a build is successful.
If a commit ID isn't specified, record it as "RELEASE" to signify that
the osbuild version is the one in the distro repos.
If the osbuild version and commit ID don't match, schedule the config
for a build.

This makes sure that changes in osbuild which don't affect the manifest
get retested.  This can catch bugs that weren't caught in osbuild and is
useful for testing internal osbuild functionality changes.

This commit also fixes a typo in the build info.json:
obuild-version -> osbuild-version