anaconda: switch container-installer to fixed ks
Use a fixed kickstart to cut corners, this needs a PoC as soon as
possible.

We need manual partitioning because we need to set the boot label on the
`/boot` partition. It is possible that this won't be necessary in the
future: https://github.com/rhinstaller/anaconda/pull/5363

We'll remove the kickstart literal once the PR is merged and the options
used are merged into the kickstart stage.