cmdutil: new function for generating a random seed
A new function generates a random seed unless the
OSBUILD_TESTING_RNG_SEED env var is set, in which case it returns the
value parsed as an integer.

The function is defined in a new internal package that will be used to
hold functions common to command line tools like gen-manifests and
build.

Co-Authored-By: Michael Vogt <michael.vogt@gmail.com>