distro/rhel9: add support for azure-sap-rhui
We don't have a snapshot yet, so the test repository is just
download.eng.bos.redhat.com for now. Let's fix it when we have a snapshot.

Regarding firewalld and libcanberra-gtk2:

The 1st party image has firewalld, but doesn't have libcanberra-gtk2. This is
actually backwards that what is in out EC2 SAP images, so I had to do some
shuffling to keep the behaviour in EC2, but have it different for Azure.